Switch back to 1.4.x of JupyterHub
We were running on a fairly recent, unreleased version of
JupyterHub to get access to
https://github.com/jupyterhub/jupyterhub/pull/3488.
It provides a seamless experience for users logging in to
data8.datahub.berkeley.edu, as it uses datahub.berkeley.edu as
the auth provider. However, it had also brought in a bunch of
unreleased changes that broke the idle culler
(https://github.com/berkeley-dsep-infra/datahub/issues/2452) -
probably around pagination.

https://github.com/jupyterhub/jupyterhub/pull/3579 backports
just the change we want to the last released version of JupyterHub,
and should hopefully fix culling.